lottery-simulator/backend/requirements.txt
# FastAPI和相关依赖
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 数据库相关
sqlalchemy>=2.0.23
alembic>=1.13.1

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 数据分析和机器学习
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
scipy>=1.11.4

# HTTP客户端
httpx>=0.25.2
requests>=2.31.0

# 工具库
python-dateutil>=2.8.2
pytz>=2023.3

# 开发和测试
pytest>=7.4.3
pytest-asyncio>=0.21.1

# 环境变量管理
python-dotenv>=1.0.0

# 日志
loguru>=0.7.2
celery==5.3.4